Extends pg with tables from database
* Extends repo based on vitaly-t's demo
  vitaly-t/pg-promise-demo
