+++ db/index.js
@@ -1,45 +1,28 @@
 'use strict';
 
 const path = require('path');
-const pgp = require('pg-promise')();
-const cn = process.env.DATABASE_URL || 'postgres://localhost/project3_dev';
-const db = pgp(cn);
 const sql = require(path.join(__dirname, 'sql'));
 
-module.exports = {
-  users: {
-    create: function(){
-      return db.none(sql.users.create);
-    },
-    // Expects {username:, password_digest:}
-    // Returns: {user_id:, username: }
-    get: function(username){
-      return db.oneOrNone(sql.users.get, username);
-    },
-    // Expects {username:, password_digest:}
-    // Returns {user_id: , username: }
-    add: function(values){
-      return db.one(sql.users.add, values);
-    }
-  },
-  hunts: {
-    create: function(){
-      return db.none(sql.hunts.create);
-    }
-  },
-  clues: {
-    create: function(){
-      return db.none(sql.clues.create);
-    }
-  },
-  participants: {
-    create: function(){
-      return db.none(sql.participants.create);
-    }
-  },
-  clues_users: {
-    create: function(){
-      return db.none(sql.clues_users.create);
-    }
+const repos = {
+  users: require(path.join(__dirname, 'repos/users')),
+  clues: require(path.join(__dirname, 'repos/clues')),
+  hunts: require(path.join(__dirname, 'repos/hunts')),
+  participants: require(path.join(__dirname, 'repos/participants')),
+  clues_users: require(path.join(__dirname, 'repos/clues_users'))
+};
+
+const opts = {
+  extend: function(obj){
+    obj.users = repos.users(obj);
+    obj.clues = repos.clues(obj);
+    obj.hunts = repos.hunts(obj);
+    obj.participants = repos.participants(obj);
+    obj.clues_users = repos.clues_users(obj);
   }
 };
+
+const pgp = require('pg-promise')(opts);
+const cn = process.env.DATABASE_URL || 'postgres://localhost/project3_dev';
+const db = pgp(cn);
+
+module.exports = db;
