+++ db/repos/clues.js
@@ -1,0 +1,12 @@
+'use strict';
+
+const path = require('path');
+const sql = require(path.join(__dirname, '../sql'));
+
+module.exports = function(db) {
+  return {
+    create: function(){
+      return db.none(sql.clues.create);
+    }
+  };
+};
