+++ db/repos/clues_users.js
@@ -1,0 +1,12 @@
+'use strict';
+
+const path = require('path');
+const sql = require(path.join(__dirname, '../sql'));
+
+module.exports = function(db) {
+  return {
+    create: function(){
+      return db.none(sql.clues_users.create);
+    }
+  };
+};
